% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outrigger.R
\name{outrigger}
\alias{outrigger}
\title{Performs outrigger regression at a point x}
\usage{
outrigger(
  formula,
  data,
  xtest,
  degree = 0,
  bandwidth = NULL,
  kernel = "epan",
  lambda = NULL,
  folds = 5,
  scoregranality = 1,
  score_df = 10
)
}
\arguments{
\item{formula}{formula (as in e.g. \code{lm}).}

\item{data}{data frame.}

\item{xtest}{datapoint(s) for local estimation.}

\item{degree}{degree of polynomial.}

\item{bandwidth}{bandwidth parameter. If NULL a rule-of-thumb bandwidth is selected by least-squares cross-validation for the standard local polynomial estimator.}

\item{kernel}{kernel for local polyomial. Default is epanechnikov kernel.}

\item{lambda}{orthogonalisation parameter. By default takes \eqn{\lambda = 4(1 + 0.5 \log(n))}.}

\item{folds}{number of folds for cross-fitting. Default is 5.}

\item{scoregranality}{the number of bins to split covariate-space \eqn{|X-x|\leq\lambda h} for conditional score estimation.}

\item{score_df}{number of degrees of freedom for score matching splines used for conditional score estimation. Input can be a numeric value (e.g. 10 on larger datasets, 6 on smaller datasets) or "cross-validation" or "cross-validation-quick", in which case score-matching-CV will be performed (may be computationally costly).}
}
\value{
If x is a single point, return a list containing:
  \describe{
    \item{\code{prediction}}{The outrigger local polynomial estimator \eqn{\hat{f}(x)} at \eqn{x}.}
    \item{\code{fitted_vector}}{For the local linear outrigger estimator, the full fitted vector at \eqn{x}}
    \item{\code{standardlocpol_fitted_vector}}{The fitted vector for the standard local polynomial}
    \item{\code{score_plot_metadata}}{Data used for \code{score_plotting} to plot the fitted conditional score functions}
  }
  If x is a vector of points, return a dataset with points x and associated outrigger fitted values.
}
\description{
Performs outrigger regression at a point x
}
