% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.outrigger_results}
\alias{print.outrigger_results}
\title{Print outrigger results}
\usage{
\method{print}{outrigger_results}(x, ...)
}
\arguments{
\item{x}{An object of class \code{outrigger_results}.}

\item{...}{Additional arguments passed to print methods.}
}
\value{
Invisibly returns the input object.
}
\description{
Print outrigger results
}
