% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/score_plotting.R
\name{score_plots}
\alias{score_plots}
\title{Plots the fitted score functions learnt in an \code{outrigger} fit.}
\usage{
score_plots(fitted_outrig, plot_together = TRUE)
}
\arguments{
\item{fitted_outrig}{a fitted object from \code{outrigger}}

\item{plot_together}{a logical denoting whether all score function estimators (across all covariate bins) should be plotted toghether. Default is TRUE.}
}
\description{
Plots the fitted score functions learnt in an \code{outrigger} fit.
}
