% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{summary.outrigger_results}
\alias{summary.outrigger_results}
\title{Summarize outrigger results}
\usage{
\method{summary}{outrigger_results}(object, ...)
}
\arguments{
\item{object}{An object of class \code{outrigger_results}.}

\item{...}{Additional arguments (unused).}
}
\value{
A list with summary statistics.
}
\description{
Summarize outrigger results
}
